Change architecture
Instead sending messages to all opened tabs now popup page is listening
for events from content script. This is a vice versa changes - content
scripts no longer listening for events, it sends request to the popup
when clicking and waiting for response.

The background page is needed because popup page can not respond when
its closed.